requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.44
botocore==1.42.44
certifi==2026.1.4
click==8.3.1
dnspython==2.8.0
fastapi==0.68.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
motor==3.7.1
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
py-ocpi==0.3.1
pydantic==1.10.26
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
rfc3986==1.5.0
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.14.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
